refactor: address review feedback on usage trend feature
- Make Usage Trend day bars keyboard-focusable (buttons with focus/blur readout) and convert the inline sparkline trigger to a semantic button.
- Extract parse_bar_chart_line out of parse_lines in the plugin runtime.
- Move the ccusage and sparkline tests into dedicated spec files.
