+++ plugins/codex/plugin.test.js
@@ -400,62 +400,6 @@ describe("codex plugin", () => {
     }
   })
 
-  it("adds model percentage text lines and a usage chart from codex ccusage", async () => {
-    const ctx = makeCtx()
-    ctx.host.fs.writeText("~/.codex/auth.json", JSON.stringify({
-      tokens: { access_token: "token" },
-      last_refresh: new Date().toISOString(),
-    }))
-    ctx.host.http.request.mockReturnValue({
-      status: 200,
-      headers: { "x-codex-primary-used-percent": "10" },
-      bodyText: JSON.stringify({}),
-    })
-    ctx.host.ccusage.query.mockReturnValue({
-      status: "ok",
-      data: {
-        daily: [
-          {
-            date: "2026-02-02",
-            totalTokens: 300,
-            models: {
-              "gpt-5.5": { totalTokens: 200 },
-              "gpt-5": { totalTokens: 100 },
-            },
-          },
-          {
-            date: "2026-02-01",
-            totalTokens: 150,
-            models: {
-              "gpt-5": { inputTokens: 30, cachedInputTokens: 20, outputTokens: 50 },
-            },
-          },
-        ],
-      },
-    })
-
-    const plugin = await loadPlugin()
-    const result = plugin.probe(ctx)
-
-    const chart = result.lines.find((line) => line.label === "Usage Trend")
-    expect(chart).toMatchObject({
-      type: "barChart",
-      note: "Estimated from local Codex logs for the selected account.",
-    })
-    expect(chart.points.map((point) => point.value)).toEqual([150, 300])
-
-    const gpt55 = result.lines.find((line) => line.label === "gpt-5.5")
-    const gpt5 = result.lines.find((line) => line.label === "gpt-5")
-    expect(gpt55).toMatchObject({
-      type: "text",
-      value: "50%",
-    })
-    expect(gpt5).toMatchObject({
-      type: "text",
-      value: "50%",
-    })
-  })
-
   it("passes CODEX_HOME to ccusage via homePath", async () => {
     const ctx = makeCtx()
     ctx.host.env.get.mockImplementation((name) => (name === "CODEX_HOME" ? "/tmp/codex-home" : null))
